change url convention in preparation for backbone
In this case we are making a second request to the server to fetch the data through ajax, but ideally the base page would be small and/or cached and provide a loading indicator. This can actually improve user experience for large files.

Backbone routers work much better when there are no changes in the base url and all changes are in the hash

The only possible negative effect here is the reduction of the users' perception of security that came from the separation between the hash and the query. However, mega.co.nz does it this way too and I haven't heard complaints.